refactor(provenance): reorganize modal into Modal/ subfolder, split cards.tsx
DatasetProvenanceModal and its section/nav/header helpers now live under
Provenance/Modal/ instead of flat alongside unrelated page-level and catalogue
components. cards.tsx (382 lines, 8 unrelated components) is split into one
file per component under Modal/cards/, with shared type helpers extracted to
Modal/cards/helpers.ts.
