+++ src/js/components/Provenance/Modal/SpatialCoverageSection.tsx
