+++ src/js/components/Provenance/Modal/SummarySection.tsx
