+++ src/js/components/Provenance/Modal/cards/index.ts
@@ -1,0 +1,8 @@
+export { CopyButton } from './CopyButton';
+export { OntologyChip } from './OntologyChip';
+export { SectionHead } from './SectionHead';
+export { LinkTile } from './LinkTile';
+export { PersonCard } from './PersonCard';
+export { PublicationCard } from './PublicationCard';
+export { FundingCard } from './FundingCard';
+export { LicenseTile } from './LicenseTile';
